pyproject.toml
[project]
name = "ohara"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.31.0",
    "httpx[socks]>=0.28.1",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "pytesseract>=0.3.13",
    "rich>=14.1.0",
    "sqlmodel>=0.0.24",
]
